Converts url params to hooks for ServiceDetails component
