+++ x-pack/plugins/apm/public/components/app/ServiceDetails/ServiceDetailTabs.tsx
@@ -12,54 +12,56 @@ import { HistoryTabs } from '../../shared/HistoryTabs';
 import { ErrorGroupOverview } from '../ErrorGroupOverview';
 import { TransactionOverview } from '../TransactionOverview';
 import { ServiceMetrics } from './ServiceMetrics';
+import { useLocation } from '../../../hooks/useLocation';
 
-interface TabsProps {
+interface Props {
   transactionTypes: string[];
   urlParams: IUrlParams;
-  location: Location;
   isRumAgent?: boolean;
 }
 
-export class ServiceDetailTabs extends React.Component<TabsProps> {
-  public render() {
-    const { transactionTypes, urlParams, location, isRumAgent } = this.props;
-    const { serviceName } = urlParams;
-    const headTransactionType = transactionTypes[0];
-    const transactionsTab = {
-      name: i18n.translate('xpack.apm.serviceDetails.transactionsTabLabel', {
-        defaultMessage: 'Transactions'
-      }),
-      path: headTransactionType
-        ? `/${serviceName}/transactions/${headTransactionType}`
-        : `/${serviceName}/transactions`,
-      routePath: `/${serviceName}/transactions/:transactionType?`,
-      render: () => (
-        <TransactionOverview
-          urlParams={urlParams}
-          serviceTransactionTypes={transactionTypes}
-        />
-      )
-    };
-    const errorsTab = {
-      name: i18n.translate('xpack.apm.serviceDetails.errorsTabLabel', {
-        defaultMessage: 'Errors'
-      }),
-      path: `/${serviceName}/errors`,
-      render: () => {
-        return <ErrorGroupOverview urlParams={urlParams} location={location} />;
-      }
-    };
-    const metricsTab = {
-      name: i18n.translate('xpack.apm.serviceDetails.metricsTabLabel', {
-        defaultMessage: 'Metrics'
-      }),
-      path: `/${serviceName}/metrics`,
-      render: () => <ServiceMetrics urlParams={urlParams} location={location} />
-    };
-    const tabs = isRumAgent
-      ? [transactionsTab, errorsTab]
-      : [transactionsTab, errorsTab, metricsTab];
+export function ServiceDetailTabs({
+  transactionTypes,
+  urlParams,
+  isRumAgent
+}: Props) {
+  const location = useLocation();
+  const { serviceName } = urlParams;
+  const headTransactionType = transactionTypes[0];
+  const transactionsTab = {
+    name: i18n.translate('xpack.apm.serviceDetails.transactionsTabLabel', {
+      defaultMessage: 'Transactions'
+    }),
+    path: headTransactionType
+      ? `/${serviceName}/transactions/${headTransactionType}`
+      : `/${serviceName}/transactions`,
+    routePath: `/${serviceName}/transactions/:transactionType?`,
+    render: () => (
+      <TransactionOverview
+        urlParams={urlParams}
+        serviceTransactionTypes={transactionTypes}
+      />
+    )
+  };
+  const errorsTab = {
+    name: i18n.translate('xpack.apm.serviceDetails.errorsTabLabel', {
+      defaultMessage: 'Errors'
+    }),
+    path: `/${serviceName}/errors`,
+    render: () => {
+      return <ErrorGroupOverview urlParams={urlParams} location={location} />;
+    }
+  };
+  const metricsTab = {
+    name: i18n.translate('xpack.apm.serviceDetails.metricsTabLabel', {
+      defaultMessage: 'Metrics'
+    }),
+    path: `/${serviceName}/metrics`,
+    render: () => <ServiceMetrics urlParams={urlParams} location={location} />
+  };
+  const tabs = isRumAgent
+    ? [transactionsTab, errorsTab]
+    : [transactionsTab, errorsTab, metricsTab];
 
-    return <HistoryTabs tabs={tabs} />;
-  }
+  return <HistoryTabs tabs={tabs} />;
 }
