+++ x-pack/plugins/apm/public/components/app/ServiceDetails/index.tsx
@@ -4,17 +4,55 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import { connect } from 'react-redux';
-import { IReduxState } from '../../../store/rootReducer';
-import { getUrlParams } from '../../../context/UrlParamsContext';
-import { ServiceDetailsView } from './view';
+import { EuiFlexGroup, EuiFlexItem, EuiSpacer, EuiTitle } from '@elastic/eui';
+import React from 'react';
+import { useFetcher } from '../../../hooks/useFetcher';
+import { loadServiceDetails } from '../../../services/rest/apm/services';
+import { FilterBar } from '../../shared/FilterBar';
+import { ServiceDetailTabs } from './ServiceDetailTabs';
+import { ServiceIntegrations } from './ServiceIntegrations';
+import { isRumAgentName } from '../../../../common/agent_name';
+import { useUrlParams } from '../../../hooks/useUrlParams';
 
-function mapStateToProps(state = {} as IReduxState) {
-  return {
-    urlParams: getUrlParams(state)
-  };
+export function ServiceDetails() {
+  const { urlParams } = useUrlParams();
+  const { serviceName, start, end, kuery } = urlParams;
+  const { data: serviceDetailsData } = useFetcher(
+    () => loadServiceDetails({ serviceName, start, end, kuery }),
+    [serviceName, start, end, kuery]
+  );
+
+  if (!serviceDetailsData) {
+    return null;
+  }
+
+  const isRumAgent = isRumAgentName(serviceDetailsData.agentName || '');
+
+  return (
+    <React.Fragment>
+      <EuiFlexGroup justifyContent="spaceBetween">
+        <EuiFlexItem>
+          <EuiTitle size="l">
+            <h1>{urlParams.serviceName}</h1>
+          </EuiTitle>
+        </EuiFlexItem>
+        <EuiFlexItem grow={false}>
+          <ServiceIntegrations
+            transactionTypes={serviceDetailsData.types}
+            urlParams={urlParams}
+          />
+        </EuiFlexItem>
+      </EuiFlexGroup>
+
+      <EuiSpacer />
+
+      <FilterBar />
+
+      <ServiceDetailTabs
+        urlParams={urlParams}
+        transactionTypes={serviceDetailsData.types}
+        isRumAgent={isRumAgent}
+      />
+    </React.Fragment>
+  );
 }
-
-const ServiceDetails = connect(mapStateToProps)(ServiceDetailsView);
-
-export { ServiceDetails };
